case/requirements.txt
Django==4.2
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
djangorestframework-simplejwt==5.2.2
